+++ main.py
@@ -1,4 +1,4 @@
-from visualization.server import create_server
+from src.visualization.server import create_server
 
 if __name__ == "__main__":
     server = create_server()
