+++ src/agents/obstacle.py
@@ -1,0 +1,12 @@
+# src/agents/obstacle.py
+from mesa import Agent
+
+
+class Obstacle(Agent):
+    """Obstacle agent. Just to add obstacles to the grid."""
+
+    def __init__(self, unique_id, model):
+        super().__init__(unique_id, model)
+
+    def step(self):
+        pass
