+++ src/agents/road.py
@@ -1,7 +1,13 @@
+# src/agenas/road.py
 from mesa import Agent
 
 
 class Road(Agent):
-    def __init__(self, unique_id, model):
+    """Road agent. Determines where the cars can move, and in which direction."""
+
+    def __init__(self, unique_id, model, direction="Left"):
         super().__init__(unique_id, model)
-        self.type = "road"  # Can be "road" or "boundary"
+        self.direction = direction
+
+    def step(self):
+        pass
